requirements.txt
PyMySQL == 0.9.3
SQLAlchemy == 1.3.16
xlwt == 1.3.0
python-dotenv == 0.13.0